requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
gunicorn>=23.0,<24.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
twilio>=9.0,<10.0
websockets>=12.0,<16.0
resend>=0.8,<1.0
email-validator>=2.2,<3.0
python-multipart>=0.0.9,<1.0
jinja2>=3.1,<4.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.2,<4.0
cryptography>=42.0,<44.0
